requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.2
transformers~=4.51.3
torch~=2.7.0
torchvision~=0.22.0
pillow~=11.2.1